refactor(chat): move terminal.output cap helper to claudette lib crate
Copilot's follow-up review on PR #949 flagged a bidirectional
coupling: `commands::env` was reaching into
`commands::chat::send::background_tasks` for the shared
`append_terminal_output_sync` helper, while `chat::send` already
depended on `commands::env` for env-provider resolve helpers. Two
command modules pointing at each other is an obvious smell, and
the file growing the helper (`background_tasks.rs`, 1268 lines) is
on the god-file watchlist in CLAUDE.md.

Lift the cap and rotation machinery into `claudette::agent::background`
where `workspace_terminal_output_path` already lives. That is the
natural home — the cap is a property of the workspace terminal file
itself, not of any specific Tauri command. Both writer paths
(`commands::chat::send::background_tasks::append_agent_bash_output`
and `commands::env::WorkspaceTerminalFileSink`) now import the
helper from the lib crate; the chat ↔ env import is gone.

The move also handles a third Copilot finding: the prior writer did
not enforce the cap when a single incoming payload was *itself*
bigger than `TERMINAL_OUTPUT_TAIL_BYTES`. Rotating then appending
still left the file at `banner + tail_cap + bytes.len()`. The
lib-crate helper now detects that regime and writes only the tail
of the payload (preceded by the truncation banner), so the
post-write file size is bounded at `banner + tail_cap` regardless
of how big the incoming append was.

While moving, also update the `workspace_terminal_output_path` doc
comment that previously promised "full history of everything that
ever ran in the workspace" — that contract is now bounded; recent
scrollback survives and older history is dropped with a visible
banner. (User-facing docs were already updated in eb0ec592.)

Tests:
- The cap unit tests move with the code into the lib crate's test
  module — they were testing lib-level behavior all along.
- New regression `append_caps_oversized_single_payload_to_tail_of_bytes`
  builds a 2× tail-sized payload, hands it to the helper in one shot,
  and asserts (a) the truncation banner is at the start, (b) the
  file body is exactly `tail_cap` bytes, and (c) only the tail half
  of the payload (the `T` bytes) survives — the head (the `H` bytes)
  is dropped.
- 19/19 lib tests pass; 12/12 tauri-side background_tasks tests
  still pass (`mirror_background_task_output_dedupes_duplicate_spawns`
  stays Tauri-side since it depends on the tokio runtime).
